pyproject.toml
[project]
name = "lng-bot1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.1",
    "cerebras-cloud-sdk>=1.67.0",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "flask>=3.1.3",
    "ipykernel>=7.2.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.21",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "langgraph-cli[inmem]>=0.4.19",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
    "wikipedia>=1.4.0",
]
